#1718: Create SWOT Deliverable Module

+++ backend/IMIS/IMIS.Application/SWOTAnalysisModule/ISWOTAnalysisService.cs
@@ -8,7 +8,7 @@ namespace IMIS.Application.SWOTAnalysisModule
     {
         Task<SWOTAnalysisDto?> GetByIdAsync(int id, CancellationToken cancellationToken);
         Task<List<SWOTAnalysisDto>?> GetByUserIdAsync(string userId, CancellationToken cancellationToken);
-        public Task<DtoPageList<SWOTAnalysisDto, SWOTAnalysis, int>> GetPaginatedAsync(int page, int pageSize, CancellationToken cancellationToken);
+        public Task<DtoPageList<SWOTAnalysisDto, SWOTAnalysis, long>> GetPaginatedAsync(int page, int pageSize, CancellationToken cancellationToken);
         Task<bool> SoftDeleteAsync(int id, CancellationToken cancellationToken);
         Task<List<SWOTAnalysisDto>?> FilterByYearAsync(int year, int noOfResults, CancellationToken cancellationToken);
     }
